+++ src/main/java/com/portfolio/service/QuestionService.java
@@ -8,17 +8,17 @@ import com.portfolio.domain.SearchVO;
 import java.util.List;
 
 public interface QuestionService {
-    long create(QuestionVO questionVO);
-    QuestionVO select(long quesNo);
+    Integer create(QuestionVO questionVO);
+    QuestionVO select(Integer quesNo);
     void update(QuestionVO questionVO);
-    void delete(long postNo);
-    long selectLastQuesNo();
+    void delete(Integer postNo);
+    Integer selectLastQuesNo();
     List<QuestionVO> selectList(QuestionVO questionVO);
     List<QAPairVO> selectPairList(QuestionVO questionVO);
     List<CategoryVO> selectCategories();
     List<QuestionVO> searchQuestions(SearchVO searchVO);
     List<QAPairVO> searchPairs(SearchVO searchVO);
-    long countQuestions(QuestionVO questionVO);
+    Integer countQuestions(QuestionVO questionVO);
 
 
 }
